+++ source/Irrlicht/COpenGLDriver.cpp
@@ -1931,7 +1931,7 @@ COpenGLDriver::SAuxContext::COpenGLVAO::~COpenGLVAO()
 void COpenGLDriver::SAuxContext::COpenGLVAO::bindBuffers(   const COpenGLBuffer* indexBuf,
                                                             const COpenGLBuffer* const* attribBufs,
                                                             const size_t offsets[asset::EVAI_COUNT],
-                                                            const size_t strides[asset::EVAI_COUNT])
+                                                            const uint32_t strides[asset::EVAI_COUNT])
 {
     uint64_t beginStamp = CNullDriver::ReallocationCounter;
